feat: vim.ui.input, avoid metatable in config
The vim.ui.input changes include shell completion and highlighting. I'll
have to add a switch for highlighting, because it doesn't work well
enough. I've also added a new function to the API, which allows the user
to have all the benefits of the new changes (in the 4th branch that
can't be reached with just the main function) without having to define
all of the input logic themselves.